+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.1 (2021-07-08)
+------------------
+
+* Generate command more fleshed out
+* Added more detailed pytests 
+
+
 0.1.0 (2021-06-30)
 ------------------